feat: add AI 😈
Okay it is still very basic, can not even choose dialogue options. But it would be weird to put that as the commit message on a later commit when it actually gets pathfinding and stuff since the AI would already sort have been in the game before that commit. And I really want to do ts as a commit message because it is so funny 😹.