feat: make bots move to dialogue locations
However, they stutter sometimes.

Also, they can not pathfind through the road obby to the store; this will likely require a pathfinding link.

Lastly, it would be optimal for them to move to the dialogue location as the players move there. Currently, they only move to the location when the dialogue starts. They are late and pulling up to the after party. This is unacceptable.