feat: finish the game as much as possible
It would have been lovely to commit all of this separately, but I was grinding the game as much as possible. It was due approximately 40 minutes ago. Do not worry, I submitted it in time. I think commiting this after the due date is fine though since the repository is not actually required at all for the game jam.

Here is everything that this commit contains, in the order that I see changes in files from top to bottom, with the files being ordered in the Visual Studio Code Staged Changes menu, which I do not know how it is ordered:

* feat: add waiting for players dialogue

Untested.

* feat: make brainrot character always use emoji

* fix: fix spelling

(If this was its own commit it would make sense, but as this is referring to "- 5 Shopping List**s**".)

* feat: add truss to obby

* feat: add sale dialogue

* feat: fade to black for cinematic camera

* feat: add cinematic camera

* fix: remove debug prints

* feat: flip pathfinding links

* chore: remove testing data

* fix: wait before collecting shopping lists

* fix: remove train barrier

* fix: remove unnecessary `task.spawn`

* style: reorder code to be more chronological

* fix: anchor some map stuff